♻️(frontend) remove i18n routing
At first we though it was a good idea but in pratice we realized it was
causing too much pain to maintain. The ultime issue was that if the URL
were not 100% translated for a given language then the dashboard was
crashing due to the fact that we cannot set child route that has a different
starting portion than its parent. And this was happening in the following
example:

Parent: /teacher/cours/:courseId/produits/:courseProductRelationId
Child:  /teacher/courses/:courseId/products/:courseProductRelationId/learners

Resulting error: An absolute child route path must start with the combined
path of all its parent routes.