💄(frontend) improve order group message color and layout
Currently, the order group message in CourseProductItem has no color set. That
mean we are not able to customize it from our factory that is weird. We decide
to use the base color of the CourseProductItem. Furthermore, the spacing around
the message is pretty weird so we propose an improvement of margin.