✨(frontend) add link to product order in syllabus
When a product that has been purchased is on going on the
syllabus page, we added a link to the order details page
of the learner dashboard.
Also, to display the "enrolled" tag of a joanie's product, we need
to find one active enrollment between order's target's enrollment.

resolve #2361
resolve #2362