⬆️(backend) upgrade to Django 4.0
Since Django 4.0, the use of `pytz` is deprecated within Django
and is replaced by `zoneinfo` to use `datetime.timezone` instead.
We drop the support of postgres version less than 12. We have
also updated the template tags `{% ifequal %}` in favor of `{% if %}`
template tag. The setting configuration `USE_L10N` is deprecated
in the settings if the value is `True`.