💄(front) rework credit card logos
Previously the credit card logos were rendered on top of a blue
background, but as it is not the case in the sale tunnel, we got
white logos on top of white background, which was not ideal. Now
the approach is to make sure logos are self-visible on their own.