🐛(back) fully define STORAGES settings
In order to use django 4.2 we have to fully define the settings
STORAGES. We can't simply override the default or the static key. By
fully defining it, the deprecated settings DEFAULT_FILE_STORAGE and
STATICFILES_STORAGE are defined by django once setup by using the one
defined in STORAGES. And the django filer correctly use them.