✨(frontend) retrieve estimated payment schedule in sale tunnel
In further development, we will stop to sell training with one shot payment.
Instead we will schedule payments. From frontend, that means, before the user
purchase a training, it should be able to know what is the planned schedule. A
first step is to retrieve the estimated payment schedule and display it in the
sale tunnel.