♻️(frontend) do not abort when validated order polling limit is reached
Currently, on payment success, when the payment polling limit to get validated
order is reached, we abort the payment. That is weird as in this use case, the
payment has succeeded, the problem is just that the notification has not been
received in time by joanie to validate the order. So instead to abort payment,
we just display a message to explain to the user that the order validation is
taking more time than expected.