✨(frontend) create SaleTunnelNotValidated step
Currently, after payment success, if we reach the poll limit, we only display
an error message to explain to the user it has to wait as its order
validation its too long. It works but it's good for UX as user has to close
the payment modal at the payment step. That's why we decided to create a
SaleTunnelNotValidated. This component is shown when the poll limit is reach
and it helps the user to understand that it has to wait a little bit to start
its training.