💄(frontend) move SaleTunnelSponsors into the left column of SaleTunnel
We are preparing the SaleTunnel to display payment schedule. Currently we are
displaying training organizations below the purchase button. With the payment
schedule, the course plan column looks a little bit empty and the layout looks
unbalanced. So in order to improve layout, we decide to move SaleTunnelSponsors
component into the left column