🔧(frontend) upgrade typescript target to es2018
With the latest upgrade of typescript, a type error is raised about the use of
regex with capturing group. Indeed, this feature is available with es2018, and
we are currently using es6 as typescript target. But currently our browser
support list allow us to upgrade the typescript target from es6 to es2018.

https://caniuse.com/?search=es2018