🔨(docker) fix first bootstrap can't connect to DB
Fix error `dockerflow.health.E001` with the message of
`Could not connect to database`.
This only occurs from a fresh installation or on the
first boot of the database.
Probably this only occurs when the dev machine has
a slower IO.
Split docker compose file, per supported database
and add a dependency of the app to db, if it's healthy,
directly on docker compose level.
Also fixes running tests on mysql locally.

fixes #2455