🐛(ci) fix npm development version
setuptools_scm generates a development version which can not be
compatible with the semver syntax used by npm. For a development
version, the version and prerelease tag should be seperate with a dash,
setuptools_scm use a period as separation. We have to transform the
version with one compatible with NPM.