✨(frontend) display checkbox to waive withdrawal right if needed
We want to prevent a user to subscribe to a training without waiving withdrawal
right if the training will start before the end of the withdrawal period. So
according the property `has_withdrawal_period` of a product, we ask to user to
check a box to explicitly waive its withdrawal right.