🔧(frontend) do not compile react-intl anymore
Previously, react-intl package was written in CommonJS and was not compiled so
we have to compile this library on our own side. Recently, our build were
broken due to a syntax issue caused by the result of the compilation of the
react-intl. After some checks, it appears this library is now written in ES
Module so we can stop to transform it and this fix our syntax issue so that's
perfect; Furthermore, we was building also `strict-uri-encode` package but it
appears this one does not exist anymore in `node_modules`