✨(plugins) added Slider plugin
This add a new CMS plugin to implement a Slider component.

The Slider is done through two plugins: SliderPlugin is the
parent container that will accept only SliderItemPlugin as
children.

This commit only include the backend, CSS and template
implementation. There is still a JavaScript component to
develop so currently the slider is static.