📌(backend) pin djangcms_link to version <5
A major version of djangocms_link has been released and it introduces breaking
changes. Those ones are currently breaking our test suites and block all our ci
pipeline. As a fast fix, we pin this dependency to unblock the situation and we
will have to work on this upgrade quickly.