📌(backend) pin all packages
Currently some packages are not pinned and this may lead to introduce bugs
during deployment that we are not able to detect during development phase
that is really weird. As richie is a library package, we cannot pin
exactly all deps but to prevent to deploy silently packages containing
breaking changes, we use a stragegy to install only minor and patch
versions of all packages. Major versions will be upgraded by renovate.