🚸(frontend) improve credit card dashboard pref view loading state
Currently, when this component was in loading state, the credit cards were not
displayed. This was weird as each time the component go to this state, the
layout flickers as the list is hidden then show again. To fix that we position
absolutely the spinner with an overlay over the view.