🐛(backend) fix header_menu template
Since dropdown menu, there are several case where we want to apply
`topbar__item--selected` class to a `topbar__item`. First the current child
has property `selected` sets to `True` then when the current child has been
configured to redirect to another page and the current page is the target of
this redirect. But in some case, the page display has an empty path and in this
case the current behavior applied the selected modified class. So in order to
fix that we need to ensure that a redirect url is sets before comparing it to
the current page url.