📌(backend) pin easy-thumbnails to 2.9.0 to avoid save JPEG error
The `bootstrap_elasticsearch` command is failing due to an error from
`easy-thumbnails` when trying to generate thumbnails for the a gif (used in
person portraits). The error was: "OSError: cannot write mode P as JPEG".

This is a known issue introduced in `easy-thumbnails` 2.10.0 (see
https://github.com/SmileyChris/easy-thumbnails/issues/653).
We pin the version to be < 2.10 as a temporary workaround.