🐛(frontend) discounted price in sales tunnel for certificate product
When we open the sales tunnel to purchase a certificate product
in a course that we are enrolled, when a discount is present
for the product offer, we would not see the discounted price.
With the latest developments on OfferingRules, we can now access
that price. When it comes to credentials, it works as usual,
and when there is no discount, the initial product price is shown.

Fix #2645