💄(frontend) handle voucher error in sale tunnel
Since we have an API rate limit in Joanie API, we want
to display the good error for the user in the sale tunnel.
If the voucher is not valid, already claimed or expired then
we display the default one. But if it's about the API rate
with 429 error code then we display the new one.